GARCA-5.0 Completar ws tipos de cuentas

+++ WebApp/GARCA.WSDATA/Managers/Interfaces/ManagerBase.cs
@@ -16,7 +16,7 @@ namespace GARCA.wsData.Managers
 
         public Task<bool> Update(Q obj);
 
-        public Task<long> Insert(Q obj);
+        public Task<int> Insert(Q obj);
 
         public Task<bool> Delete(Q obj);
